Issue 233

+++ develop/lambda/api/umesse/cm.js
@@ -11,6 +11,7 @@ const {
 const { validation } = require("umesse-lib/validation");
 const { dynamodbManager } = require("umesse-lib/utils/dynamodbManager");
 const { s3Manager } = require("umesse-lib/utils/s3Manager");
+const { sqsManager } = require("umesse-lib/utils/sqsManager");
 
 // CM取得（一覧・個別）
 exports.getCm = async (unisCustomerCd, cmId) => {
@@ -75,7 +76,7 @@ exports.createCm = async (unisCustomerCd, body) => {
     const cmId = generateId(unisCustomerCd, "c");
 
     // CM結合、S3へPUT
-    const seconds = await generateCm(unisCustomerCd, cmId, body);
+    const seconds = await generateCm(unisCustomerCd, cmId, body.materials);
     if (!seconds) throw "generate cm failed";
 
     // 署名付きURLの発行
@@ -152,16 +153,60 @@ exports.updateCm = async (unisCustomerCd, cmId, body) => {
 
     // TODO: CMステータス状態によるチェック
 
+    let res = "";
+    let dataProcessType = "";
+    let status = "";
     // CM作成中の場合
     if (cm.status == constants.cmStatus.CREATING) {
-      // TODO: add sqs
+      // sqs send
+      const params = {
+        MessageBody: JSON.stringify({
+          unisCustomerCd: unisCustomerCd,
+          cmId: cmId,
+        }),
+        QueueUrl: CONVERTER_SQS_QUEUE_URL,
+        DelaySeconds: 0,
+      };
+
+      res = await SQS.sqsManager(params);
+      if (!res) throw "update failed";
+
       cm.status = constants.cmStatus.CONVERT;
+      dataProcessType = "01";
+      status = "0";
     } else {
       // CMアップロード
       if (body.uploadSystem) {
-        // TODO:
         cm.status = constants.cmStatus.EXTERNAL_UPLOADING;
+        dataProcessType = "02";
+        status = "1";
       }
+    }
+
+    // CMアップロード
+    if (body.uploadSystem) {
+      const item = {
+        unisCustomerCd: unisCustomerCd,
+        dataProcessType: dataProcessType,
+        cmId: cmId,
+        cmName: cm.title,
+        cmCommentManuscript: cm.description,
+        startDatetime: cm.startDate,
+        endDatetime: cm.endDate,
+        productionType: cm.productionType,
+        contentTime: cm.seconds,
+        sceneCd: cm.scene.sceneCd,
+        uploadSystem: body.uploadSystem,
+        status: status,
+        timestamp: timestamp(),
+      };
+
+      res = await dynamodbManager.put(
+        constants.dynamoDbTable().external,
+        item,
+        {}
+      );
+      if (!res) throw "put failed";
     }
 
     // DynamoDBのデータ更新
@@ -179,7 +224,7 @@ exports.updateCm = async (unisCustomerCd, cmId, body) => {
     };
     debuglog(JSON.stringify({ key: key, options: options }));
 
-    const res = await dynamodbManager.update(
+    res = await dynamodbManager.update(
       constants.dynamoDbTable().users,
       key,
       options
@@ -258,25 +303,26 @@ exports.deleteCm = async (unisCustomerCd, cmId) => {
 
 // CM結合処理
 function generateCm(unisCustomerCd, cmId, materials) {
-  const isWindows = process.platform === 'win32';
-  const isLocal = process.env.environment === 'local';
+  const isWindows = process.platform === "win32";
+  const isLocal = process.env.environment === "local";
   const isLocalWindows = isLocal && isWindows;
 
   return new Promise(async function (resolve, reject) {
-    // FIXME: materials 一旦仮
-    const list = [
-      "chime/サンプル01.mp3",
-      "chime/サンプル02.mp3",
-      "bgm/サンプル01.mp3",
-      "narration/サンプル01.mp3",
-      "narration/サンプル02.mp3",
-      "narration/サンプル03.mp3",
-    ];
-    const workDir = `/tmp/${unisCustomerCd}/mix/${cmId}`;
-    const windowsWorkDir = `${process.cwd()}\\tmp\\${unisCustomerCd}\\mix\\${cmId}`; // windows only
+    const workDir = `/tmp/umesse/${unisCustomerCd}/mix/${cmId}`;
+    const windowsWorkDir = `${process.cwd()}\\tmp\\umesse\\${unisCustomerCd}\\mix\\${cmId}`; // windows only
     const ffmpeg = `ffmpeg -hide_banner`;
+    const has = {
+      startChime: "startChime" in materials,
+      endChime: "endChime" in materials,
+      bgm: "bgm" in materials,
+    };
 
     try {
+      // ナレーションチェック
+      if (!materials.narrations || materials.narrations.length < 1)
+        throw "not narration";
+
+      // workディレクトリ作成
       let initDirCommand = ``;
       if (!isLocalWindows) {
         initDirCommand = `mkdir -p ${workDir} && rm -f ${workDir}/*`;
@@ -287,32 +333,114 @@ function generateCm(unisCustomerCd, cmId, materials) {
       }
       execSync(initDirCommand);
 
-      let paths = "";
-      for (const [key, value] of list.entries()) {
-        debuglog(`key: ${key}, value: ${value}`);
-        const res = await s3Manager.get(constants.s3Bucket().contents, value);
-        if (!res || !res.Body) throw "getObject failed";
-        const filePath = !isLocalWindows ? workDir : windowsWorkDir;
-        fs.writeFileSync(`${filePath}/${key}.mp3`, res.Body);
-        paths += `-i ${workDir}/${key}.mp3 `;
+      // S3から無音ファイルを取得
+      let filePath = !isLocalWindows
+        ? `/tmp/umesse`
+        : `${process.cwd()}\\tmp\\umesse`;
+      if (!(await getContents(constants.s3Bucket().contents, "silent.mp3", filePath, "silent.mp3")))
+        throw "getObject failed";
+
+      let paths = "-i /tmp/umesse/silent.mp3 ";
+      let options = "";
+      let extra = "";
+      let index = 0;
+
+      // S3からファイル取得、コマンド作成
+      for (let [key, value] of Object.entries(materials).sort()) {
+        debuglog(JSON.stringify({ key: key, value: value }));
+        let fileName = "";
+
+        switch (key) {
+          case "narrations":
+            let count = 0;
+            let narrations = "";
+
+            for (let v of value) {
+              let target = `narration/${v.contentsId}.mp3`;
+              let bucket = constants.s3Bucket().contents;
+
+              if (v.contentsId.match(`^${unisCustomerCd}-r-[0-9a-z]{8}$`)) {
+                // レコーディング音源の場合
+                target = `${unisCustomerCd}/${constants.resourceCategory.RECORDING}/${v.contentsId}.mp3`;
+                bucket = constants.s3Bucket().users;
+              } else if (v.contentsId.match(`^${unisCustomerCd}-t-[0-9a-z]{8}$`)) {
+                // TTS音源の場合
+                target = `${unisCustomerCd}/${constants.resourceCategory.TTS}/${v.contentsId}.mp3`;
+                bucket = constants.s3Bucket().users;
+              }
+
+              fileName = `narration_${++count}`;
+              if (!(await getContents(bucket, target, workDir, `${fileName}.mp3`)))
+                throw "getObject failed";
+              paths += `-i ${workDir}/${fileName}.mp3 `;
+              narrations += `[${fileName}]`;
+              // 無音カット ボリューム調整　末尾に3秒無音追加
+              options += createCommand(["silent", "volume", "apad"], `[${++index}:a]`, v.volume, 3, fileName);
+            }
+            // ナレーションを結合
+            extra += createCommand(["concat"], narrations, null, count, "mix");
+            if (has.bgm) {
+              // BGMがある場合は、前後に3秒無音追加
+              extra += createCommand(["adelay", "apad"], "[mix]", null, 3, "mix");
+            }
+            break;
+
+          case "startChime":
+            fileName = "start_chime";
+            if (!(await getContents(constants.s3Bucket().contents, `chime/${value.contentsId}.mp3`, workDir, `${fileName}.mp3`)))
+              throw "getObject failed";
+            paths += `-i ${workDir}/${fileName}.mp3 `;
+            // 無音カット ボリューム調整　末尾に1秒無音追加
+            options += createCommand(["silent", "volume", "apad"], `[${++index}:a]`, value.volume, 1, fileName);
+            break;
+
+          case "endChime":
+            fileName = "end_chime";
+            if (!(await getContents(constants.s3Bucket().contents, `chime/${value.contentsId}.mp3`, workDir, `${fileName}.mp3`)))
+              throw "getObject failed";
+            paths += `-i ${workDir}/${fileName}.mp3 `;
+            // 無音カット ボリューム調整　頭にBGMあり3秒、BGMなし1秒無音追加
+            options += createCommand(["silent", "volume", "adelay"], `[${++index}:a]`, value.volume, has.bgm ? 3 : 1, fileName);
+            break;
+
+          case "bgm":
+            fileName = "bgm";
+            if (!(await getContents(constants.s3Bucket().contents, `bgm/${value.contentsId}.mp3`, workDir, `${fileName}.mp3`)))
+              throw "getObject failed";
+            paths += `-i ${workDir}/${fileName}.mp3 `;
+            // 無音カット ボリューム調整　BGMを無限ループ
+            options += createCommand(["silent", "volume", "aloop"], `[${++index}:a]`, value.volume, -1, fileName);
+            break;
+        }
       }
 
-      // FIXME: 各パラメーターをチューニング
-      let command = `${ffmpeg} ${paths} \
-        -filter_complex ' \
-          [0:a]volume=0.5[start_chime]; \
-          [1:a]volume=0.5,adelay=3s|3s[end_chime]; \
-          [2:a]volume=0.5,aloop=2:2.14748e+009[bgm]; \
-          [3:a]volume=3.0,adelay=3s|3s[narration1]; \
-          [4:a]volume=3.0,adelay=3s|3s[narration2]; \
-          [5:a]volume=3.0,adelay=3s|3s,apad=pad_dur=5[narration3]; \
-          [narration1][narration2][narration3]concat=n=3:v=0:a=1[join]; \
-          [join][bgm]amix=duration=shortest[mix]; \
-          [mix][end_chime]acrossfade=d=3[last]; \
-          [start_chime][last]concat=n=2:v=0:a=1 \
-        ' -y ${workDir}/${cmId}.mp3`;
+      if (has.bgm) {
+        // 結合したナレーションとBGMをMIX
+        extra += createCommand(["amix"], "[mix][bgm]", null, null, "mix");
+      }
+      if (has.startChime) {
+        // 結合したナレーションと開始チャイムを結合
+        extra += createCommand(["concat"], "[start_chime][mix]", null, 2, "mix");
+      }
+      if (has.endChime) {
+        // 結合したナレーションと終了チャイムを結合、フェードアウト3秒
+        extra += createCommand(["acrossfade"], "[mix][end_chime]", null, 3, "mix");
+      } else if (!has.endChime && has.bgm) {
+        // 無音を3秒に伸ばす
+        extra += createCommand(["adelay"], "[0:a]", null, 3, "silent");
+        // 結合したナレーションと無音を結合、フェードアウト3秒
+        extra += createCommand(["acrossfade"], "[mix][silent]", null, 3, "mix");
+      }
+      // 前後に無音を追加
+      extra += createCommand(["concat"], "[0:a][mix][0:a]", null, 3, null);
+
+      // MIX処理実施
+      let command = `${ffmpeg} ${paths} -filter_complex '${options}${extra}' -y ${workDir}/${cmId}.mp3`;
       if (isLocalWindows) {
-        fs.writeFileSync(`${windowsWorkDir}/ffmped-command`, `#!/bin/bash\n${command}`);
+        fs.writeFileSync(
+          `${windowsWorkDir}/ffmped-command`,
+          `#!/bin/bash\n${command}`
+        );
         command = `docker run --name ffmpeg-runner --rm \
           -v ${process.cwd()}\\..\\layer\\bin:/usr/local/bin \
           -v ${windowsWorkDir}:${workDir} \
@@ -322,32 +450,33 @@ function generateCm(unisCustomerCd, cmId, materials) {
       debuglog(command);
       execSync(command);
 
-      // FIXME: get seconds... 他に方法があるか
+      // CMの秒数を取得
       let secondsCommand = `${ffmpeg} -hide_banner -i ${workDir}/${cmId}.mp3 2>&1 | \
-        grep 'Duration' | cut -d ' ' -f 4 | cut -d '.' -f 1`
+        grep 'Duration' | cut -d ' ' -f 4 | cut -d '.' -f 1`;
       if (isLocalWindows) {
-        fs.writeFileSync(`${windowsWorkDir}/ffmpeg-seconds-command`, `#!/bin/bash\n${secondsCommand}`);
+        fs.writeFileSync(
+          `${windowsWorkDir}/ffmpeg-seconds-command`,
+          `#!/bin/bash\n${secondsCommand}`
+        );
         secondsCommand = `docker run --name ffmpeg-runner --rm \
           -v ${process.cwd()}\\..\\layer\\bin:/usr/local/bin \
           -v ${windowsWorkDir}:${workDir} \
           centos:latest \
           sh ${workDir}/ffmpeg-seconds-command`;
       }
-      const seconds = execSync(secondsCommand)
-        .toString()
-        .replace(/\n/g, "");
-      debuglog(`seconds: ${seconds}`);
+      const seconds = execSync(secondsCommand).toString().replace(/\n/g, "");
 
       const cmFilePath = !isLocalWindows ? workDir : windowsWorkDir;
       const fileStream = fs.createReadStream(`${cmFilePath}/${cmId}.mp3`);
       fileStream.on("error", (e) => {
         throw e;
       });
-      await s3Manager.put(
+      const res = await s3Manager.put(
         constants.s3Bucket().users,
         `users/${unisCustomerCd}/cm/${cmId}.mp3`,
         fileStream
       );
+      if (!res) throw "putObject failed";
       debuglog("generate complete");
       resolve(seconds);
     } catch (e) {
@@ -355,4 +484,42 @@ function generateCm(unisCustomerCd, cmId, materials) {
       reject();
     }
   });
+}
+
+async function getContents(bucket, target, outputPath, outputFileName) {
+  try {
+    const res = await s3Manager.get(bucket, target);
+    if (!res || !res.Body) throw "getObject failed";
+    fs.writeFileSync(`${outputPath}/${outputFileName}`, res.Body);
+    return true;
+  } catch (e) {
+    console.log(e);
+    return false;
+  }
+}
+
+function createCommand(process, index, volume, num, div) {
+  const command = {
+    silent: `silenceremove=start_periods=1:stop_periods=1:detection=peak`,
+    volume: `volume=${volume / 100}`,
+    adelay: `adelay=${num}s|${num}s`,
+    apad: `apad=pad_dur=${num}`,
+    aloop: `aloop=${num}:2.14748e+009`,
+    concat: `concat=n=${num}:v=0:a=1`,
+    amix: `amix=duration=shortest`,
+    acrossfade: `acrossfade=d=${num}`,
+  };
+
+  let res = `${index}`;
+  let options = [];
+  for (const key of process) {
+    if (key in command) {
+      options.push(command[key]);
+    }
+  }
+  res += options.join();
+  if (div) {
+    res += `[${div}];`;
+  }
+  return res;
 }
